(d) Correct handling of archived repos where user doesn't have permissions
NOTE: further investigation brought doubt on whether this change was appropriate. See the note in the next commit for
details.

Usually when the token owner doesn't have the necessary permissions, the GitHub API request for permissions returns a 403
HTTP status, and this is used as the indicator of the lack of such permissions. However, it turns out that the
combination of lack of permissions in an archived repository results in a 404 HTTP status instead. This status was
previously considered to only occur when something was wrong, and so the script was not written to be resilient to it.

There was already a check for whether the repository is archived, but this was done after the permissions API request.

A 404 status from the permissions API request in non-archived repos is still considered an unexpected error.