(d) Ignore 404 HTTP status from permissions API request
On 2021-07-12 I suddenly started getting 404 HTTP statuses from API requests for permissions. This is happening for repos that have not had any change since the previous working state.

I originally associated the problem with archived repos where I didn't have permissions (e.g., arduino/docs-hardware) and added special code to handle that (see the previous commit). But then I ran into it again with arduino/ide-test, which is not archived.

Ignoring all 404 errors is not really a good fix because there might be cases where this indicates a real problem, which
will be masked by ignoring it. However, I am wondering if this is a transient issue caused by a bug on GitHub's end. So
I'm going to wait a while before wasting more time on this, but I need the script working in the meantime, thus the hacky
workaround.