Add comments indicating source to all "template" workflows
The infrastructure of this repository is based on a collection of reusable "template" GitHub Actions workflows. Those
workflows are maintained in a centralized repository.

Documenting the upstream source of these workflows facilitates pulling advancements made to the upstream files into this
repository, and also facilitates pushing advancements made here back upstream.