Use best practices syntax in shell variable references
The shell language has a complex and obscure variable reference syntax. This can lead to bugs in cases where a variable
reference is interspersed with what was intended to be literal characters but happens to be syntax instead. For this
reason, it is best practices to wrap the variable name in braces whenever it is adjacent to literal characters.