Configure link check to use Accept-Encoding header for docs.github.com links
The link check job of the "Check Markdown" workflow had frequent intermittent spurious failures recently, caused by
links under the docs.github.com domain returning 403 HTTP status.

Others experiencing the same problem reported that they were able to work around the problem by providing a custom
`Accept-Encoding` HTTP request header. Although I was not able to find any explanation of why, it did appear to resolve
the problem.

This is a change that was made in the upstream "template" this file is based on. I don't know whether or not the
workaround is still necessary (I haven't experienced any spurious link check failures from this domain recently, but
that might be due to the workaround having been implemented in many of the repositories), but it does no harm so it is
best to sync with the upstream "template" file regardless.