Add editorconfig-checker invocation step to workflow
The editorconfig-checker/action-editorconfig-checker action is used in the "Check General Formatting" GitHub Actions
workflow.

The behavior of that action changed starting from the 2.0.0 release. Previously it ran editorconfig-checker on the files
in the workspace. The action now solely serves to make editorconfig-checker available in the runner machine for use in
subsequent steps of the job (equivalent to the various popular "setup-*" actions for other frameworks/languages/tools).

This means it is now necessary to add a new step to the workflow to actually invoke editorconfig-checker.